Move create one-one tests to new dir
No test changes, all were already using the new framework
